Revert "checkpointing current progress on controller visualization refactor"
This reverts commit 034893a2d86bbb1d3db1e2d89a83fba4fc3590a0.
